feat: implementar roles y permisos efectivos

+++ routes/Api/ActivityApi.php
@@ -1,14 +1,8 @@
 <?php
-
 use App\Http\Controllers\ActivityController;
-use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
-
-Route::group(["middleware" => ["auth:sanctum"]], function () {
-    Route::get('activity', [ActivityController::class, 'index']);
-    Route::post('activity', [ActivityController::class, 'store']);
-    Route::get('activity/{id}', [ActivityController::class, 'show']);
-    Route::put('activity/{id}', [ActivityController::class, 'update']);
-    Route::delete('activity/{id}', [ActivityController::class, 'destroy']);
-
-});
+Route::get('activity', [ActivityController::class, 'index'])->middleware('permission:content.view');
+Route::get('activity/{id}', [ActivityController::class, 'show'])->middleware('permission:content.view');
+Route::post('activity', [ActivityController::class, 'store'])->middleware('permission:content.manage');
+Route::put('activity/{id}', [ActivityController::class, 'update'])->middleware('permission:content.manage');
+Route::delete('activity/{id}', [ActivityController::class, 'destroy'])->middleware('permission:content.manage');
