+++ routes/Api/AllyApi.php
@@ -1,15 +1,8 @@
 <?php
-
 use App\Http\Controllers\AllyController;
-use App\Http\Controllers\OdsController;
-
 use Illuminate\Support\Facades\Route;
-
-Route::group(["middleware" => ["auth:sanctum"]], function () {
-    Route::get('ally', [AllyController::class, 'index']);
-    Route::post('ally', [AllyController::class, 'store']);
-    Route::get('ally/{id}', [AllyController::class, 'show']);
-    Route::post('ally/{id}', [AllyController::class, 'update']);
-    Route::delete('ally/{id}', [AllyController::class, 'destroy']);
-
-});
+Route::get('ally', [AllyController::class, 'index'])->middleware('permission:content.view');
+Route::get('ally/{id}', [AllyController::class, 'show'])->middleware('permission:content.view');
+Route::post('ally', [AllyController::class, 'store'])->middleware('permission:content.manage');
+Route::post('ally/{id}', [AllyController::class, 'update'])->middleware('permission:content.manage');
+Route::delete('ally/{id}', [AllyController::class, 'destroy'])->middleware('permission:content.manage');
