+++ routes/Api/DonationApi.php
@@ -1,14 +1,8 @@
 <?php
-
 use App\Http\Controllers\DonationController;
-use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
-
-Route::group(["middleware" => ["auth:sanctum"]], function () {
-    Route::get('donation', [DonationController::class, 'index']);
-    Route::post('donation', [DonationController::class, 'store']);
-    Route::get('donation/{id}', [DonationController::class, 'show']);
-    Route::post('donation/{id}', [DonationController::class, 'update']);
-    Route::delete('donation/{id}', [DonationController::class, 'destroy']);
-
-});
+Route::get('donation', [DonationController::class, 'index'])->middleware('permission:content.view');
+Route::get('donation/{id}', [DonationController::class, 'show'])->middleware('permission:content.view');
+Route::post('donation', [DonationController::class, 'store'])->middleware('permission:content.manage');
+Route::post('donation/{id}', [DonationController::class, 'update'])->middleware('permission:content.manage');
+Route::delete('donation/{id}', [DonationController::class, 'destroy'])->middleware('permission:content.manage');
