+++ routes/Api/SurveyQuestionApi.php
@@ -1,15 +1,10 @@
 <?php
 
 use App\Http\Controllers\SurveyQuestionController;
-use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
-Route::group(["middleware" => ["auth:sanctum"]], function () {
-    Route::get('surveyquestion', [SurveyQuestionController::class, 'index']);
-    Route::post('surveyquestion', [SurveyQuestionController::class, 'store']);
-    Route::get('surveyquestion/{id}', [SurveyQuestionController::class, 'show']);
-    Route::put('surveyquestion/{id}', [SurveyQuestionController::class, 'update']);
-    Route::delete('surveyquestion/{id}', [SurveyQuestionController::class, 'destroy']);
-
-});
-
+Route::get('surveyquestion', [SurveyQuestionController::class, 'index'])->middleware('permission:surveys.view');
+Route::get('surveyquestion/{id}', [SurveyQuestionController::class, 'show'])->middleware('permission:surveys.view');
+Route::post('surveyquestion', [SurveyQuestionController::class, 'store'])->middleware('permission:surveys.manage');
+Route::put('surveyquestion/{id}', [SurveyQuestionController::class, 'update'])->middleware('permission:surveys.manage');
+Route::delete('surveyquestion/{id}', [SurveyQuestionController::class, 'destroy'])->middleware('permission:surveys.manage');
