+++ routes/Api/SurveyQuestionOdsApi.php
@@ -3,12 +3,9 @@
 use App\Http\Controllers\Organization\SurveyQuestionOdsController;
 use Illuminate\Support\Facades\Route;
 
-Route::group(['middleware' => ['auth:sanctum']], function () {
-    Route::get('surveyquestionods', [SurveyQuestionOdsController::class, 'list']);
-    Route::get('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'show']);
-    Route::post('surveyquestionods', [SurveyQuestionOdsController::class, 'store']);
-    Route::put('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'update']);
-    Route::delete('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'destroy']);
-    Route::get('surveys/{surveyId}/charts-by-ods', [SurveyQuestionOdsController::class, 'chartsByOds']);
-
-});
+Route::get('surveyquestionods', [SurveyQuestionOdsController::class, 'list'])->middleware('permission:surveys.view');
+Route::get('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'show'])->middleware('permission:surveys.view');
+Route::get('surveys/{surveyId}/charts-by-ods', [SurveyQuestionOdsController::class, 'chartsByOds'])->middleware('permission:surveys.view');
+Route::post('surveyquestionods', [SurveyQuestionOdsController::class, 'store'])->middleware('permission:surveys.manage');
+Route::put('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'update'])->middleware('permission:surveys.manage');
+Route::delete('surveyquestionods/{id}', [SurveyQuestionOdsController::class, 'destroy'])->middleware('permission:surveys.manage');
