+++ routes/Api/SurveyQuestionOptionApi.php
@@ -3,11 +3,8 @@
 use App\Http\Controllers\SurveyQuestionOptionController;
 use Illuminate\Support\Facades\Route;
 
-Route::group(["middleware" => ["auth:sanctum"]], function () {
-    Route::get('surveyquestionoption', [SurveyQuestionOptionController::class, 'index']);
-    Route::post('surveyquestionoption', [SurveyQuestionOptionController::class, 'store']);
-    Route::get('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'show']);
-    Route::put('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'update']);
-    Route::delete('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'destroy']);
-
-});
+Route::get('surveyquestionoption', [SurveyQuestionOptionController::class, 'index'])->middleware('permission:surveys.view');
+Route::get('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'show'])->middleware('permission:surveys.view');
+Route::post('surveyquestionoption', [SurveyQuestionOptionController::class, 'store'])->middleware('permission:surveys.manage');
+Route::put('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'update'])->middleware('permission:surveys.manage');
+Route::delete('surveyquestionoption/{id}', [SurveyQuestionOptionController::class, 'destroy'])->middleware('permission:surveys.manage');
